Camera now begins at an appropriate height to the terrain chunk below it. Landscape height now able to be properly calibrated with a float in the Project1Game class.

+++ Project 2 - Terrain Loading/FractalTools.cs
@@ -14,7 +14,7 @@ namespace Project1
     class FractalTools {
 
         // Generates a large fractal before subdividing it and serialising those divisions to binary files.
-        public static void GenerateTerrainChunks(int terrainSizeFactor, int terrainRangeFactor, int chunkSizeFactor){
+        public static void GenerateTerrainChunks(int terrainSizeFactor, float terrainRangeFactor, int chunkSizeFactor){
             int numberOfChunks = (int)Math.Pow(4, terrainSizeFactor-chunkSizeFactor),
                 chunkWidth     = (int)Math.Pow(2, chunkSizeFactor)+1,
                 terrainWidth   = (int)Math.Pow(2, terrainSizeFactor)+1,
